add new version meta box 5.1.2 for learpress 3.2.6.8

+++ inc/libraries/meta-box/js/media.js
@@ -1,8 +1,4 @@
-/* global jQuery, _,i18nRwmbMedia */
-
-window.rwmb = window.rwmb || {};
-
-jQuery( function ( $ ) {
+( function ( $, wp, _, rwmb, i18n ) {
 	'use strict';
 
 	var views = rwmb.views = rwmb.views || {},
@@ -12,7 +8,9 @@ jQuery( function ( $ ) {
 		MediaCollection, Controller, MediaField, MediaList, MediaItem, MediaButton, MediaStatus, EditMedia,
 		MediaDetails, MediaLibrary, MediaSelect;
 
-	MediaCollection = models.MediaCollection = media.model.Attachments.extend( {
+	MediaCollection = Backbone.Collection.extend( {
+		model: wp.media.model.Attachment,
+
 		initialize: function ( models, options ) {
 			this.controller = options.controller || new models.Controller;
 			this.on( 'add remove reset', function () {
@@ -20,34 +18,35 @@ jQuery( function ( $ ) {
 				this.controller.set( 'length', this.length );
 				this.controller.set( 'full', max > 0 && this.length >= max );
 			} );
-
-			media.model.Attachments.prototype.initialize.call( this, models, options );
 		},
 
 		add: function ( models, options ) {
 			var max = this.controller.get( 'maxFiles' ),
 				left = max - this.length;
 
-			if ( max > 0 && left <= 0 ) {
+			if ( ! models || ( max > 0 && left <= 0 ) ) {
 				return this;
 			}
-			if( models) {
-				if ( ! models.hasOwnProperty( 'length' ) ) {
-					models = [models];
-				} else if ( models instanceof media.model.Attachments ) {
-					models = models.models;
-				}
+			if ( ! models.hasOwnProperty( 'length' ) ) {
+				models = [models];
+			} else if ( models instanceof media.model.Attachments ) {
+				models = models.models;
 			}
+
+			models = _.difference( models, this.models );
 			if ( left > 0 ) {
-				models = _.difference( models, this.models );
 				models = _.first( models, left );
 			}
 
-			return media.model.Attachments.prototype.add.call( this, models, options );
+			Backbone.Collection.prototype.add.call( this, models, options );
 		},
 
 		remove: function ( models, options ) {
-			models = media.model.Attachments.prototype.remove.call( this, models, options );
+			// Don't remove models if event is not fired from MB plugin.
+			if( ! $( event.target ).closest( '.rwmb-field, [data-class="rwmb-field"]' ).length ) {
+				return;
+			}
+			models = Backbone.Collection.prototype.remove.call( this, models, options );
 			if ( this.controller.get( 'forceDelete' ) === true ) {
 				models = ! _.isArray( models ) ? [models] : models;
 				_.each( models, function ( model ) {
@@ -92,24 +91,6 @@ jQuery( function ( $ ) {
 					this.get( 'items' ).destroyAll();
 				}
 			} );
-		},
-
-
-		// Load initial media
-		load: function () {
-			if ( _.isEmpty( this.get( 'ids' ) ) ) {
-				return;
-			}
-			this.get( 'items' ).props.set( {
-				query: true,
-				include: this.get( 'ids' ),
-				orderby: 'post__in',
-				order: 'ASC',
-				type: this.get( 'mimeType' ),
-				perPage: this.get( 'maxFiles' ) || - 1
-			} );
-			// Get more then trigger ready
-			this.get( 'items' ).more();
 		}
 	} );
 
@@ -141,11 +122,8 @@ jQuery( function ( $ ) {
 			this.createAddButton();
 			this.createStatus();
 
-			// Render
 			this.render();
-
-			// Load media
-			this.controller.load();
+			this.loadInitialAttachments();
 
 			// Listen for destroy event on input
 			this.$input.on( 'remove', function () {
@@ -161,8 +139,18 @@ jQuery( function ( $ ) {
 			}, 500 ) );
 
 			this.controller.get( 'items' ).on( 'remove', _.debounce( function () {
-				that.$input.val( '' );
+				that.$input.val( '' ).trigger( 'change' );
 			}, 500 ) );
+		},
+
+		loadInitialAttachments: function () {
+			if ( ! this.$input.val() ) {
+				return;
+			}
+			var models = this.$input.data( 'attachments' ).map( function( attachment ) {
+				return wp.media.model.Attachment.create( attachment );
+			} );
+			this.controller.get( 'items' ).add( models );
 		},
 
 		// Creates media list
@@ -215,14 +203,20 @@ jQuery( function ( $ ) {
 				},
 				function ( item ) {
 					return item.cid;
-				} );
+				}
+			);
 
 			this.listenTo( this.collection, 'add', this.addItemView );
 			this.listenTo( this.collection, 'remove', this.removeItemView );
 			this.listenTo( this.collection, 'reset', this.resetItemViews );
 
-			// Sort media using sortable
-			this.initSortable();
+			// Sort items using helper 'clone' to prevent trigger click on the image, which means reselect.
+			this.$el.sortable( {
+				helper : 'clone',
+				update: function( event, ui ) {
+					ui.item.find( rwmb.inputSelectors ).first().trigger( 'mb_change' );
+				}
+			} );
 		},
 
 		listenToItemView: function ( itemView ) {
@@ -231,19 +225,17 @@ jQuery( function ( $ ) {
 			this.listenTo( itemView, 'click:edit', this.editItem );
 		},
 
-		//Add item view
 		addItemView: function ( item ) {
 			var index = this.collection.indexOf( item ),
-				itemEl = this.getItemView( item ).el;
+				itemEl = this.getItemView( item ).el,
+				$children = this.$el.children();
 
 			if ( 0 >= index ) {
 				this.$el.prepend( itemEl );
-			}
-			else if ( this.$el.children().length <= index ) {
+			} else if ( $children.length <= index ) {
 				this.$el.append( itemEl )
-			}
-			else {
-				this.$el.children().eq( index - 1 ).after( itemEl );
+			} else {
+				$children.eq( index - 1 ).after( itemEl );
 			}
 		},
 
@@ -258,29 +250,31 @@ jQuery( function ( $ ) {
 
 		resetItemViews: function( items ){
 			var that = this;
-			_.each( that.models, function( item ){
-				 that.removeItemView( item );
-			 } );
-			items.each( function( item ) {
-				that.addItemView( item );
-			} );
+			_.each( that.models, that.removeItemView );
+			items.each( that.addItemView );
 		},
 
 		switchItem: function ( item ) {
 			if ( this._switchFrame ) {
-				//this.stopListening( this._frame );
 				this._switchFrame.dispose();
 			}
 			this._switchFrame = new MediaSelect( {
-				className: 'media-frame rwmb-media-frame',
 				multiple: false,
-				title: i18nRwmbMedia.select,
 				editing: true,
 				library: {
 					type: this.controller.get( 'mimeType' )
 				},
-				edit: this.controller.get( 'items' )
+				edit: this.collection
 			} );
+
+			// Refresh content when frame opens
+			this._switchFrame.on( 'open', function() {
+				var frameContent = this._switchFrame.content.get();
+				if ( frameContent && frameContent.collection ) {
+					frameContent.collection.mirroring._hasMore = true;
+					frameContent.collection.more();
+				}
+			}, this );
 
 			this._switchFrame.on( 'select', function () {
 				var selection = this._switchFrame.state().get( 'selection' ),
@@ -298,60 +292,21 @@ jQuery( function ( $ ) {
 		},
 
 		editItem: function ( item ) {
-			// Destroy the previous collection frame.
 			if ( this._editFrame ) {
-				//this.stopListening( this._frame );
 				this._editFrame.dispose();
 			}
 
-			// Trigger the media frame to open the correct item
+			// Trigger the media frame to open the correct item.
 			this._editFrame = new EditMedia( {
 				frame: 'edit-attachments',
 				controller: {
-					// Needed to trick Edit modal to think there is a gridRouter
-					gridRouter: {
-						navigate: function ( destination ) {
-						},
-						baseUrl: function ( url ) {
-						}
-					}
+					gridRouter: new wp.media.view.MediaFrame.Manage.Router()
 				},
 				library: this.collection,
 				model: item
 			} );
 
 			this._editFrame.open();
-		},
-
-		initSortable: function () {
-			var collection = this.controller.get( 'items' );
-			this.$el.sortable( {
-				// Clone the element and the clone will be dragged. Prevent trigger click on the image, which means reselect.
-				helper : 'clone',
-
-				// Record the initial `index` of the dragged model.
-				start: function ( event, ui ) {
-					ui.item.data( 'sortableIndexStart', ui.item.index() );
-				},
-
-				// Update the model's index in the collection.
-				// Do so silently, as the view is already accurate.
-				update: function ( event, ui ) {
-					var model = collection.at( ui.item.data( 'sortableIndexStart' ) );
-
-					// Silently shift the model to its new index.
-					collection.remove( model, {
-						silent: true
-					} );
-					collection.add( model, {
-						silent: true,
-						at: ui.item.index()
-					} );
-
-					// Fire the `reset` event to ensure other collections sync.
-					collection.trigger( 'reset', collection );
-				}
-			} );
 		}
 	} );
 
@@ -382,8 +337,7 @@ jQuery( function ( $ ) {
 		},
 
 		render: function () {
-			var attributes = _.clone( this.controller.attributes );
-			this.$el.html( this.template( attributes ) );
+			this.$el.html( this.template( this.controller.toJSON() ) );
 		}
 	} );
 
@@ -397,33 +351,38 @@ jQuery( function ( $ ) {
 		template: wp.template( 'rwmb-media-button' ),
 		events: {
 			'click .button': function () {
-				// Destroy the previous collection frame.
 				if ( this._frame ) {
-					//this.stopListening( this._frame );
 					this._frame.dispose();
 				}
 				var maxFiles = this.controller.get( 'maxFiles' );
 				this._frame = new MediaSelect( {
-					className: 'media-frame rwmb-media-frame',
 					multiple: maxFiles > 1 || maxFiles <= 0 ? 'add' : false,
-					title: i18nRwmbMedia.select,
 					editing: true,
 					library: {
 						type: this.controller.get( 'mimeType' )
 					},
-					edit: this.controller.get( 'items' )
+					edit: this.collection
 				} );
+
+				// Refresh content when frame opens
+				this._frame.on( 'open', function() {
+					var frameContent = this._frame.content.get();
+					if ( frameContent && frameContent.collection ) {
+						frameContent.collection.mirroring._hasMore = true;
+						frameContent.collection.more();
+					}
+				}, this );
 
 				this._frame.on( 'select', function () {
 					var selection = this._frame.state().get( 'selection' );
-					this.controller.get( 'items' ).add( selection.models );
+					this.collection.add( selection.models );
 				}, this );
 
 				this._frame.open();
 			}
 		},
 		render: function () {
-			this.$el.html( this.template( {text: i18nRwmbMedia.add} ) );
+			this.$el.html( this.template( {text: i18n.add} ) );
 			return this;
 		},
 
@@ -450,37 +409,32 @@ jQuery( function ( $ ) {
 		template: wp.template( 'rwmb-media-item' ),
 		initialize: function ( options ) {
 			this.controller = options.controller;
+			this.collection = this.controller.get( 'items' );
 			this.render();
-			this.listenTo( this.model, 'change', function () {
-				this.render();
-			} );
+			this.listenTo( this.model, 'change', this.render );
 
 			this.$el.data( 'id', this.model.cid );
 		},
 
-
 		events: {
-			'click .rwmb-overlay': function () {
+			'click .rwmb-image-overlay': function ( e ) {
+				e.preventDefault();
 				this.trigger( 'click:switch', this.model );
-				return false;
 			},
-
-			// Event when remove button clicked
-			'click .rwmb-remove-media': function () {
+			'click .rwmb-remove-media': function ( e ) {
+				e.preventDefault();
 				this.trigger( 'click:remove', this.model );
-				return false;
 			},
-
-			'click .rwmb-edit-media': function () {
+			'click .rwmb-edit-media': function ( e ) {
+				e.preventDefault();
 				this.trigger( 'click:edit', this.model );
-				return false;
 			}
 		},
 
 		render: function () {
-			var attrs = _.clone( this.model.attributes );
-			attrs.controller = _.clone( this.controller.attributes );
-			this.$el.html( this.template( attrs ) );
+			var data = this.model.toJSON();
+			data.controller = this.controller.toJSON();
+			this.$el.html( this.template( data ) );
 			return this;
 		}
 	} );
@@ -515,7 +469,7 @@ jQuery( function ( $ ) {
 	MediaLibrary = media.controller.Library.extend( {
 		defaults: _.defaults( {
 			multiple: 'add',
-			filterable: 'uploaded',
+			filterable: 'all',
 			priority: 100,
 			syncSelection: false
 		}, media.controller.Library.prototype.defaults ),
@@ -556,6 +510,9 @@ jQuery( function ( $ ) {
 		createStates: function () {
 			var options = this.options;
 
+			// Add reference so we know MediaFrame belongs to MB plugin.
+			this.$el.attr( 'data-class', 'rwmb-field' );
+
 			if ( this.options.states ) {
 				return;
 			}
@@ -566,7 +523,6 @@ jQuery( function ( $ ) {
 				new MediaLibrary( {
 					library: media.query( options.library ),
 					multiple: options.multiple,
-					title: options.title,
 					priority: 20
 				} )
 			] );
@@ -598,20 +554,30 @@ jQuery( function ( $ ) {
 				controller: this,
 				model: this.model
 			} ) );
-		}
+		},
+		resetRoute: function() {}
 	} );
 
-	/**
-	 * Initialize media fields
-	 * @return void
-	 */
 	function initMediaField() {
-		var view = new MediaField( { input: this } );
-		//Remove old then add new
-		$( this ).siblings( 'div.rwmb-media-view' ).remove();
-		$( this ).after( view.el );
+		var $this = $( this ),
+			view = $this.data( 'view' );
+
+		if ( view ) {
+			return;
+		}
+
+		view = new MediaField( { input: this } );
+
+		$this.siblings( '.rwmb-media-view' ).remove();
+		$this.after( view.el );
+		$this.data( 'view', view );
 	}
 
-	$( '.rwmb-file_advanced' ).each( initMediaField );
-	$( document ).on( 'clone', '.rwmb-file_advanced', initMediaField );
-} );
+	function init( e ) {
+		$( e.target ).find( '.rwmb-file_advanced' ).each( initMediaField );
+	}
+
+	rwmb.$document
+		.on( 'mb_ready', init )
+		.on( 'clone', '.rwmb-file_advanced', initMediaField );
+} )( jQuery, wp, _, rwmb, i18nRwmbMedia );
